+++ apiphany-core/src/main/java/org/apiphany/security/keys/KeyExchangeHandler.java
@@ -1,4 +1,4 @@
-package org.apiphany.security;
+package org.apiphany.security.keys;
 
 import java.security.KeyPair;
 import java.security.PrivateKey;